Fix first build and startup crash of the WinUI app
- Add repo nuget.config to bypass user-level package source mapping
- Upgrade WindowsAppSDK 1.6 -> 1.8 (1.6 XamlCompiler breaks silently on SDK 10)
- Align Interop code with CsWin32 0.3.106 API surface (W-suffix stripping,
  ComTypes.FILETIME, SafeHandle overloads)
- Rework window-level bindings to x:Bind functions; drop unused BoolToVisibilityConverter
- Make brush converters DependencyObjects: {ThemeResource} only assigns to
  dependency properties, so plain CLR props crashed App.xaml parsing at launch

+++ src/TaskManager.App/Interop/FileTimeExtensions.cs
@@ -1,4 +1,4 @@
-using Windows.Win32.Foundation;
+using System.Runtime.InteropServices.ComTypes;
 
 namespace TaskManager.App.Interop;
 
@@ -9,6 +9,8 @@ namespace TaskManager.App.Interop;
 /// </summary>
 internal static class FileTimeExtensions
 {
+    // FILETIME's fields are signed int32; go through uint so the low half's sign bit
+    // never bleeds into the combined value.
     public static ulong ToUInt64(this FILETIME value) =>
-        ((ulong)value.dwHighDateTime << 32) | value.dwLowDateTime;
+        ((ulong)(uint)value.dwHighDateTime << 32) | (uint)value.dwLowDateTime;
 }
